Changes to the Server database to support coverage tracking.  These changes are probably not backwards compatible, meaning it will be necessary for users of the new server, to have the new version of POSIT mobile.
The gps_sample table has a new field for "time", which is the
time the point was retrieved on the phone.  The functions used
to send points from the mobile to the server and functions used
to retrieve points for displaying on the server are also changed.
The 'time' attribute is used to order the points so the track
can be properly reconstructed. :